This commit adds new server status variables ("show status like 'Ndb%'"). These are: * sorted_scan_count * pushed_queries_defined * pushed_queries_dropped * pushed_queries_executed * pushed_reads (See ha_ndbcluster.h for an explanation of what these do.)
I addition, pruned_scan_count will now also count pushed pruned scans. But
this is hard to test, since there is an existing error that prevents scans
pushed by the handler from being pruned.

This commit also removes a couple of warnings due to size_z being assigned to
Uint32.

+++ storage/ndb/src/ndbapi/NdbQueryBuilder.cpp
@@ -1092,7 +1092,7 @@ int NdbDoubleConstOperandImpl::convertDouble()
 int NdbCharConstOperandImpl::convertChar()
 {
   Uint32 len = m_column->getLength();
-  Uint32 srclen = (m_value) ? strlen(m_value) : 0;
+  Uint32 srclen = (m_value) ? static_cast<Uint32>(strlen(m_value)) : 0;
   if (unlikely(srclen > len)) {
     // TODO: Truncates: May silently remove trailing spaces:
     return QRY_CHAR_OPERAND_TRUNCATED;
@@ -1114,7 +1114,7 @@ int NdbCharConstOperandImpl::convertChar()
 int NdbCharConstOperandImpl::convertVChar()
 {
   Uint32 maxlen = m_column->getLength();
-  Uint32 len = (m_value) ? strlen(m_value) : 0;
+  Uint32 len = (m_value) ? static_cast<Uint32>(strlen(m_value)) : 0;
   if (unlikely(len > maxlen)) {
     // TODO: Truncates: May silently remove trailing spaces:
     return QRY_CHAR_OPERAND_TRUNCATED;
